requirements.txt
torch>=2.0.0
transformers>=4.51.0
datasets>=2.14.0
huggingface-hub>=0.17.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0
sentence-transformers>=2.7.0 